backend/requirements.txt
fastapi
uvicorn[standard]
aiohttp
sqlalchemy
psycopg2-binary
pgvector
sentence-transformers
python-multipart
numpy
alembic
python-dotenv
pillow
PyPDF2
python-docx
textract
pdfplumber
pytesseract
